requirements.txt
Django==1.5.2
Pygments==1.6
South==0.8.2
argparse==1.2.1
bpython==0.12
django-extensions==1.2.0
django-model-utils==1.5.0
mock==1.0.1
model-mommy==1.1
psycopg2==2.5.1
six==1.4.1
wsgiref==0.1.2
